+++ MediaBrowserTv/app/src/main/java/tv/emby/embyatv/playback/IPlaybackOverlayFragment.java
@@ -6,11 +6,11 @@ import mediabrowser.model.dto.BaseItemDto;
  * Created by Eric on 4/28/2015.
  */
 public interface IPlaybackOverlayFragment {
-    public void setCurrentTime(int time);
+    public void setCurrentTime(long time);
     public void setFadingEnabled(boolean value);
     public void setPlayPauseActionState(int state);
     public void updateDisplay();
-    public void updateEndTime(int timeLeft);
+    public void updateEndTime(long timeLeft);
     public void removeQueueItem(int pos);
     public void nextItemThresholdHit(BaseItemDto nextItem);
     public void finish();
